Move integration tests to subfolder
